master

+++ muduo_se/net/EventLoop.h
@@ -25,7 +25,7 @@ class EventLoop{
         void removeChannel(Channel* c);
         timeId runat(const timestamp1& t,const TimerCallBack& f);
         timeId runafter(int delay,const TimerCallBack& f);
-        timeId runevery(int interval,const TimerCallBack& f);
+        timeId runevery(double interval,const TimerCallBack& f);
         void cancel(timeId id);
         static EventLoop* instance();
     private:
